iommu/ummu: Support UMMU iopf capability
ubmmu driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ID40QK

-----------------------------------------------------------

This patch add the iopf (IO Page Fault) queue management. In the SVA
scenario (hardware and process sharing a page table), the iopf processing
is enabled based on the device capability. When the iopf occurs, the
interrupt processing is performed and the result is delivered to the
hardware after the page is supplemented.

Signed-off-by: Liubin Bin <shuliubin@huawei.com>
Signed-off-by: Chao Gao <gaochao24@huawei.com>
Signed-off-by: Liming An <anliming1@h-partners.com>