ub: ubase: Fix the issue of mismatch between the ubase aeqe structure and the protocol.
drivers inclusion
category: Bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/ID5SNV
CVE: NA

------------------------------------------------------------------

For ubase_aeqe structure, the field of num is 32 bits, but the protocol
specifies that this field should be 20 bits.
This patch modifies the bit width of the num field to be consistent
with the protocol.

Fixes: 9cc0f22b5d69 ("ub: ubase: support for async event process")
Signed-off-by: Zhang Lei <zhanglei917@huawei.com>